requirements.txt
anyio==3.6.2
attrs==23.1.0
cachetools==5.3.3
certifi==2023.7.22
cffi==1.17.1
cfgv==3.3.1
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
coverage==7.0.0
cryptography==44.0.0
distlib==0.3.6
exceptiongroup==1.0.4
fastapi==0.110.1
filelock==3.8.0
google-auth==2.32.0
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
identify==2.5.9
idna==3.4
iniconfig==1.1.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
nodeenv==1.7.0
numpy==1.26.2
orjson==3.8.6
packaging==21.3
pandas==1.5.2
platformdirs==2.5.4
pluggy==1.0.0
pre-commit==2.20.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==1.10.2
PyJWT==2.10.1
pyparsing==3.0.9
pytest==7.2.0
pytest-asyncio==0.23.7
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0
referencing==0.31.1
requests==2.31.0
rfc3986==1.5.0
rpds-py==0.13.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.37.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.0
uvicorn==0.20.0
virtualenv==20.16.7
